Delete ConnectionUtils in favor of Postgrex.Utils
